Add support for TLS over SMTP
Prior to this commit we would only support SMTP SSL, which is not the
same as TLS over SMTP. Some SMTP servers implement only one of the
two options.

By first trying SMTP SSL and if that fails try to do TLS over SMTP
we still support both ways of sending email over an encrypted
connection.